ICS V8.60 - Part 14
OverbyteIcsFileCopy should now build on OSX, albeit with some warnings, not tested.  Also builds on Delphi 7. 

Added new units to C++ 10.2 and 10.3 packages, unfortunately appears to be impossible to add new DCR files to the C++ packages through the IDE, get duplicate error, so icons are missing from these new units and some earlier ones for VCL, should be OK for FMX.  

Barring bug fixes and more documentation, V8.60 should now be complete, build and install on Delphi 7 to 10.3 and C++ 10.2 and 10.3.  Please try it. 



git-svn-id: http://svn.overbyte.be:8443/svn/ics/trunk@1413 17a045a4-f01d-714f-988e-c09671ad6f7c